WIP - Use a null object for the build cache service handle
